blog_project/requirements.txt
asgiref==3.9.1
Django==5.0.6
django-ckeditor==6.7.1
django-js-asset==3.1.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.8.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
